==> 1_Report.ipynb <==
from ydata_profiling import ProfileReport
import pandas as pd

1. Import

df = pd.read_csv('data/cars_infos.csv')

2. Reporting

profile = ProfileReport(df, title='CSV_profile_report')
profile.to_file('data/csv_profile_report.html')

<p style='color:grey;'>📝 Un overview html est désormais disponible dans le dossier data</p>
