==> 2_Nettoyage.ipynb <==
import pandas as pd

### 1. Import

df = pd.read_csv('data/cars_infos.csv')
df.columns

### 2. Traitement des colonnes

a. Indexion :<br>
- car_ID

# on supprimme la colonne et on garde l'index du dataframe
df = df.drop('car_ID', axis=1)

b. Séparation :  
- CarName

# on extrait le modèle inclut dans le CarName
df['Modèle'] = df.CarName.apply(lambda x : ' '.join(x.split(' ')[1:]).title())

# on ne garde que la marque dans le CarName
df.CarName = df.CarName.apply(lambda x : x.split(' ')[0].title())

# on renomme
df = df.rename(columns={'CarName':'Marque'})

df[['Marque','Modèle']].sample(5)

c. Traduction :  
- fueltype, carbody, drivewheel, enginelocation

# on définit les dictionnaires

fueltype_fr = {
    'gas' : 'Essence',
    'diesel' : 'Diesel'
}

carbody_fr = {
    'convertible':'Cabriolet',
    'hatchback':'Hayon',
    'sedan':'Berline',
    'wagon':'Break',
    'hardtop':'Coupé'
}

drivewheel_fr = {
    'rwd' : 'Propulsion arrière',
    'fwd' : 'Traction avant',
    '4wd' : 'Quatre roues motrices' 
}

enginelocation_fr = {
    'front':'Avant',
    'rear':'Arrière'
}

# On mappe les traductions grâce aux dictionnaires ci-dessus:
df.fueltype = df.fueltype.apply(lambda x : fueltype_fr[x])
df.carbody = df.carbody.apply(lambda x : carbody_fr[x])
df.drivewheel = df.drivewheel.apply(lambda x: drivewheel_fr[x])
df.enginelocation = df.enginelocation.apply(lambda x: enginelocation_fr[x])

# On traduit les colonnes
df = df.rename(columns={
    'fueltype':'Carburant',
    'carbody':'Classe',
    'drivewheel':'Transmission',
    'enginelocation':'PositionMoteur'
    })

df[['Carburant', 'Classe','Transmission','PositionMoteur']].sample(5)

d. Correspondance:  
- aspiration

#La colonne indique si la voiture possède un turbo
df.aspiration.unique()

#On transforme la colonne en booléen Turbo = True/False
aspiration_fr = {
    'std' : False,
    'turbo' : True,
}
df.aspiration = df.aspiration.apply(lambda x : aspiration_fr[x])
df = df.rename(columns={'aspiration':'Turbo'})
df[['Turbo']].sample(5)

e. Numérisation:  
- doornumber, cylindernumber

# On définit le dictionnaire
number_fr = {
    'two' : 2,
    'three' : 3,
    'four' : 4,
    'five' : 5,
    'six' : 6,
    'eight' : 8,
    'twelve' : 12
}

#On transforme les valeurs textuelles en numérique grâce au dictionnaire:

df.doornumber = df.doornumber.apply(lambda x: number_fr[x]+1)
# /!\ On ajoutte un car en Français on compte le coffre comme porte

df.cylindernumber = df.cylindernumber.apply(lambda x: number_fr[x])

#on traduit les colonnes
df = df.rename(columns={
    'doornumber':'Portes',
    'cylindernumber':'Cylindres'})
df[['Portes','Cylindres']].sample(5)

f. Conversion:  
- wheelbase, carlenght, carheight, curbweight, enginesize, boreratio, stroke, citympg, highwaympg

#on convertit de pouces(in) en centimètres(cm)
df.wheelbase = df.wheelbase.apply(lambda x: round(2.54*x))
df.carlength = df.carlength.apply(lambda x: round(2.54*x))
df.carwidth = df.carwidth.apply(lambda x: round(2.54*x))
df.carheight = df.carheight.apply(lambda x: round(2.54*x))
df.stroke = df.stroke.apply(lambda x: round(2.54*x))

#On convertit de livres(lbs) en tonnes(T)
df.curbweight = df.curbweight.apply(lambda x: round(0.45359*x)/1000)

#On convertit de pouces cubes(in³) en centimètres cubes(cm³)
df.enginesize = df.enginesize.apply(lambda x: round(16.3871*x))

#On convertit de pouces(in) en milimètres
df.boreratio = df.boreratio.apply(lambda x : round(25.4*x))

#On convertit de Miles per gallon(mpg) en litres au 100(L/100km)
df.citympg = df.citympg.apply(lambda x: round(x/235.214,3))
df.highwaympg = df.highwaympg.apply(lambda x: round(x/235.214,3))

#on traduit les colonnes en indiquant l'unité de mesure métrique
df = df.rename(columns={
    'wheelbase':'Empattement(cm)',
    'carlength':'Longueur(cm)',
    'carheight':'Hauteur(cm)',
    'carwidth' : 'Largeur(cm)',
    'curbweight' : 'Poids(t)',
    'enginesize': 'Moteur(cm³)',
    'boreratio' : 'Alésage(mm)',
    'stroke' : 'Piston(cm)',
    'citympg' : 'ConsommationVille(L/100km)',
    'highwaympg' : 'ConsommationAutoroute(L/100km)'
    })
df[['Empattement(cm)', 'Longueur(cm)', 'Hauteur(cm)', 'Poids(t)', 'Alésage(mm)', 'Piston(cm)', 'ConsommationVille(L/100km)', 'ConsommationAutoroute(L/100km)']].sample(5)

g. Mise en forme:  
- enginetype, fuelsystem

# Le type de moteur et le système d'injection est representé par des codes suffisament éloquents
enginetype = df.enginetype.unique()
fuelsystem = df.fuelsystem.unique()
print(f'Engines types : {enginetype} \n Fuel systems : {fuelsystem}')

# On mets les codes en lettres capitales
df.enginetype = df.enginetype.apply(lambda x: x.upper())
df.fuelsystem = df.fuelsystem.apply(lambda x: x.upper())

# On renomme
df = df.rename(columns={
    'enginetype':'TypeMoteur',
    'fuelsystem':'Injecteur'
    })

df[['TypeMoteur','Injecteur']].sample(5)

h. Nom de colonnes:
- symboling, compressionratio, horsepower, peakrpm

# On traduit les colonnes restantes
df = df.rename(columns={
    'symboling' : 'CoteRisque',
    'compressionratio' : 'TauxCompression',
    'horsepower' : 'Chevaux',
    'peakrpm' : 'Régime(tr/min)',
    'price' : 'Prix($)',
})

i. Repositionnement des colonnes

# On reposition les colonnes en les regroupant par catégories
df = df[[
    'Prix($)', #Notre target
    'Marque', 'Modèle', 'Classe', 'Portes', #Caracteristiques physiques
    'Longueur(cm)', 'Hauteur(cm)', 'Largeur(cm)','Empattement(cm)', 'Moteur(cm³)','Poids(t)', #Dimensions
    'Carburant', 'ConsommationVille(L/100km)', 'ConsommationAutoroute(L/100km)', #Consommation
    'PositionMoteur','Transmission', #Mécanique
    'Chevaux', 'Régime(tr/min)', 'Turbo', 'Cylindres', #Puissance
    'Injecteur','TypeMoteur', #Codes
    'Alésage(mm)','Piston(cm)', #Pièces
    'TauxCompression', #Ratio
    'CoteRisque' #Assurance
    ]]

### 3. Traitement des lignes

a. Marque

# Certaines marques ont été mal orthographiées
df.Marque.unique()

# On corrige en appliquant le bon orthographe sur des masques
df.loc[df.Marque == 'Alfa-Romero', 'Marque'] = 'Alfa-Romeo'
df.loc[df.Marque == 'Bmw', 'Marque'] = 'BMW'
df.loc[df.Marque == 'Toyouta', 'Marque'] = 'Toyota'
df.loc[df.Marque == 'Maxda', 'Marque'] = 'Mazda'
df.loc[df.Marque == 'Porcshce', 'Marque'] = 'Porsche'
df.loc[(df.Marque == 'Vokswagen') | (df.Marque == 'Vw'), 'Marque'] = 'Volkswagen'

# Toutes les marques sont orthographiées pareil
df.Marque.unique()

b. Modèle

# On mets le même orthographe au modèle orthographié "100Ls" et "100 Ls"
df.loc[df.Modèle == '100 Ls', 'Modèle'] = '100Ls'
df.Modèle.unique()

### 4. Export


df.to_csv('data/cars_infos_cleaned.csv', index=False)
